elementsd-tests: consolidate setup code
